+++ setup.py
@@ -23,7 +23,8 @@ install_requirements = [
     'PyYAML==4.2b4',
     'web3==4.5.0',
     'ocean-secret-store-client==0.0.1',
-    'requests'
+    'requests',
+    'deprecated',
     # web3 requires eth-abi, requests, and more,
     # so those will be installed too.
     # See https://github.com/ethereum/web3.py/blob/master/setup.py
