+++ squid_py/agreements/templates.py
@@ -1,0 +1,10 @@
+
+class OceanAgreementTemplates:
+    def __init__(self, keeper, asset_resolver, asset_consumer, config):
+        self._keeper = keeper
+        self._asset_resolver = asset_resolver
+        self._asset_consumer = asset_consumer
+        self._config = config
+
+    def create(self, template_id):
+        pass
