+++ squid_py/ocean/__init__.py
@@ -1,3 +1,1 @@
-from .account import Account
 from .ocean import Ocean
-from .ocean_base import OceanBase
