--- squid_py/ocean/ocean_base.py
@@ -1,13 +1,0 @@
-"""Ocean module."""
-
-
-class OceanBase:
-    """Ocean abstract class."""
-
-    def __init__(self, _id):
-        """
-        Initialize the OceanBase class.
-
-        :param _id: Id, str
-        """
-        self.id = _id
