+++ tests/assets/test_assets.py
